Add files via upload
A java program that asks the user to enter a URL. The programs connects to the URL, and writes the webpage data to a file named "WebPage.html". The program assumes that the URL points to only text data, and other data formats such as multimedia etc. won't be written.